build: Port from MANIFEST.in to setuptools_scm
Using `setuptools_scm` allows us to reliably ensure that all files in
git are in the sdist, which avoids problems caused by occasionally
forgetting to update `MANIFEST.in` when adding a new directory.

`MANIFEST.in` still needs to be kept (in a reduced form) because there
are a few files needed by webpack and by kolibri-installer-android which
are only created at dist time, and exist outside setuptools.

References:
 - https://github.com/pypa/setuptools_scm#readme
 - https://pip.pypa.io/en/stable/reference/build-system/pyproject-toml/
 - https://github.com/pypa/setuptools_scm/issues/190
 - https://packaging.python.org/en/latest/guides/using-manifest-in/

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://github.com/endlessm/kolibri-explore-plugin/issues/647