build: Generate package version number automatically from git
`setuptools_scm` can do this for us, so that makes things a little
easier to maintain. It now generates `_version.py` for us automatically.

The `__version__` symbol from that file still has to be imported into
`kolibri_explore_plugin/__init__.py` so that it’s re-exported from the
`kolibri_explore_plugin` module. Although none of the code in this
repository needs it, the `webpack_json.py` code from kolibri.git
automatically looks at the `__version__` symbol of all the plugin
modules it loads, and errors if it’s not present.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>