NavBar: Increase spacing between Feedback icon and label
To make it consistent with the other elements on the navigation bar, we
use the same weird fact that when the labels are on a line by themselves
they are rendered with some extra space chars around them, and that
makes the spacing between the icons and labels look good.

Make at some point in the future we should make that alignment more
explicit, but for now I'm just adding a comment in the explaining what
happens and stating the requirement.

https://github.com/endlessm/kolibri-explore-plugin/issues/667