Add custom channel and content node APIs without caching
Upstream's `ChannelMetadataViewSet` and `ContentNodeViewset` use a
caching scheme to avoid unnecessary database queries. Unfortunately, the
django side of the cache doesn't respect request `Cache-Control` headers
and will only perform database queries if the cache entry has expired or
the URL has changed. The client side works around this by including the
`contentCacheKey` query parameter, but the value used is only updated
when the page is reloaded.

Since the client side has no way to know when the content has been
updated, the only way to ensure it receives fresh data about the content
is to not use any caching. This commit adds downstream viewsets that
don't use upstream's `metadata_cache` helper. Instead, a `Cache-Control`
value `no-cache` is be returned. This tells the browser that it can
cache the response, but it needs to revalidate with the server every
time.

All is not lost, though. Since responses use the content cache key as
the `ETag` and browsers will include an `If-None-Match` header with the
previous `ETag` value, the server can still respond with a 304 Not
Modified and skip the database query if the content cache key hasn't
changed.