packages: Unify browserslists used by packages
The root package already uses `browserslist-config-kolibri`, so remove
the manually maintained `.browserslistrc` files and specify
`browserslist-config-kolibri` in the package configuration for all
packages too.

As with the previous commits, the compatibility level is implicitly set
by Kolibri anyway (its JS runs in the same context as the JS from
kolibri-explore-plugin, and it all has to work), so there’s no point in
having a different browserlist from it.

Signed-off-by: Philip Withnall <philip@tecnocode.co.uk>

Helps: https://github.com/endlessm/kolibri-installer-android/issues/159