Add @babel/preset-env to dependencies
It’s needed for the babel support we use. Ideally, we’d pull it in by
adding a `devDependency` on `kolibri-tools`, but I can’t find a way to
depend on the development version of Kolibri, and it needs to be tightly
coupled to the development version of kolibri-explore-plugin so that
they use the same version of node.js.

So, the next best thing is to add a dependency on the bit of Babel which
we use. This was done by running `yarn add -W '@babel/preset-env'`.

Signed-off-by: Philip Withnall <philip@tecnocode.co.uk>