Fix bootstrap arguments passed to pipenv
If you don't specify any options to bootstrap.sh, `pipenv_install_args`
will retain its default of an empty string and pipenv will be passed an
empty quoted argument. That's wrong and some versions of pipenv fail
expecting the argument to be a package requirement.

Instead, initialize the variable as an array and append to it as needed.
Bash expands an empty array to nothing using the `"${var[@]}"` notation.