Override remotecontentimport task priority to high
When we enqueue the `remotecontentimport` task from
`CollectionDownloadManager`, we can specify the priority of the jobs as
high. However, when a user initiates a content download from the
frontend using the tasks API, that's not possible. Since the upstream
default priority is regular, it can get blocked behind any other queued
tasks such as the background thumbnail downloads.

This adds a copy of upstream's `remotecontentimport` with the default
priority set to high and arranges for it to be used throughout the
plugin.

Helps: #739