Factor out job running code
This separates the helpers for managing jobs from the task definitions
and the views using them. This should allow for cleaner reuse. There
should be no functional changes.

One important change here is that the task types are initially stored as
callable paths. This will allow the task data to be serialized later
without immediately resolving the task functions.