Handle incomplete background jobs with StorageHook
Using the Django app `ready` method wasn't the right approach. Those run
whenever Django loads the app and before the database models are
prepared. Instead, use a Kolibri `StorageHook`, which allows immediately
restarting the jobs. Previously jobs would only be restarted when the
Kolibri was restarted.

While here, include `CANCELED` jobs in the restart logic. We're not
cancelling these jobs, but the worker will cancel running jobs when it
stops. Since the storage hook runs even when the worker is shutting
down, this will requeue an incomplete job for when the worker next
starts.