Add python testing infrastructure
This adds the infrastructure for testing the Python backend with pytest.
A custom pytest plugin is needed to control Kolibri and Django
initialization. Tests are expected to use the helpers from pytest-django
to interact with Kolibri and Django. Actual tests will be added in a
subsequent commit.

Note that this also declares `kolibri` as a pipenv dependency as it's
required to be present for testing. The existing `requests` dependency
is removed as Kolibri's vendored copy will always be used at runtime.