build: Update to vue-cli 5 and consolidate on webpack 5
This is the only version of vue-cli which supports webpack 5; the
version we were using only supports webpack 4. kolibri-tools and the
rest of Kolibri assume webpack 5, which led to both versions being
imported, and webpack 5 plugins being used by a webpack 4 compilation
process (spawned by vue-cli-service), which threw exceptions everywhere.

So, standardise on webpack 5, and hence update to vue-cli 5.

Signed-off-by: Philip Withnall <philip@tecnocode.co.uk>

Helps: #771