loading-screen: Set up i18n
We can’t use Kolibri’s internationalisation functions, as loading-screen
is standalone and executed outside the context of a Kolibri plugin. So
we use formatjs and its Vue integration.

See the new `README.md` file for full details.

I decided not to add support for `babel-plugin-formatjs` as it
duplicates checks already done by the `formatjs` CLI and would
cause the loading-screen babel setup to diverge from the babel setup for
the rest of the plugin.

References:
 - https://formatjs.io/docs/getting-started/application-workflow
 - https://formatjs.io/docs/intl/#formatmessage
 - https://formatjs.io/docs/vue-intl/#tooling
 - https://formatjs.io/docs/tooling/babel-plugin/
 - https://webpack.js.org/api/module-methods/
 - https://webpack.js.org/guides/code-splitting/#dynamic-imports

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #770