Don't run build:i18n from build script
This is currently breaking the release because `i18n-extract-backend`
always causes the `en.po` file to be updated. This in turn causes
`setuptools_scm` to create a dev version of the wheel, which is promptly
rejected by PyPI. For now, drop `build:i18n` from the generic `build`
script until a better fix can be found.

Fixes: #787