build: Rework kolibri-tools commands for extracting translation messages
It turns out that kolibri-tools only accepts one
`--namespace`/`--searchPath` argument pair on the command line, and
silently ignores all of them except the last one that you provide.

This means the previous command was not extracting translatable messages
for ek-components or template-ui, just the main plugin and the
welcome-screen (which was the last pair of arguments).

Rearrange the commands so that each plugin/package is extracted
separately. This works for message extraction, but unfortunately each
invocation of `kolibri-tools` currently deletes *all* CSV files in the
target directory, so we’re left with only the final CSV file.

That’s fixed by https://github.com/learningequality/kolibri/pull/11226.
In the meantime, whoever periodically regenerates the CSV/JSON files
will need to do so with that patch applied locally.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>