Change routing and redirects for Welcome
The root URL now redirects to the Welcome page if there isn't a download
ongoing and the content is empty. Or to the Download page if there is a
download ongoing. A new decideWelcome handler is used for these
redirects.

The decideDownload handler is still kept to start/resume the download
when the Download page loads.