Build standalone packages with --no-module
If you target older browsers in vue-cli 5, it will build both a legacy
single file webpack bundle as well as a modern modular webpack
bundle[1]. Unfortunately, that grows the output by nearly 50%. Since
each of the channel overrides built from template-ui and loading-screen
are standalone web apps distributed as a local zip files rather than
downloaded from a remote server, we want to keep the size down as much
as possible. Use the `--no-module` option so that only the legacy
webpack bundle is built.

If we want to drop IE11 support in the future, then vue-cli will
automatically build only the modern modular webpack bundle. Particularly
for loading-screen, the only consumer is the Endless Key apps where we
can be reasonably sure the webview is modern.

1. https://cli.vuejs.org/migrations/migrate-from-v4.html#the-build-command-and-modern-mode