kolibriApi: Expose createTranslator from kolibri.utils.i18n
So that it can be used by template-ui and ek-components without
depending on kolibri.utils.i18n themselves. This is needed because they
are used from Kolibri apps which don’t already depend on kolibri-tools,
and hence would grow significantly in size if they had to depend on it.

The naming of this method is important, as `kolibri-tools
i18n-extract-messages` parses the JS AST for function calls to
`createTranslator` with these arguments in order to work out which
string literals need to be translated.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>