i18n: Extract backend messages from all locales
It appears that upstream Kolibri only updates the `en` Django PO file
and uploads that to crowdin, which then updates all the other locale PO
files on its service. That's not going to work for us, so just extract
the backend messages for all locales to keep the PO files up to date.

That generates PO files for all the locales Kolibri supports, which we
have no intention of supporting right now. That's harmless, though.