ek-components: Delay calling createTranslator() until actually needed
`createTranslator()` in ek-components now relies on `window.kolibri`
being available, and that’s only initialised as part of the app setup,
which is after all the imports are processed.

This causes the `createTranslator()` calls to fail at import time as
`window.kolibri` is undefined.

Fix that by wrapping them in an internal function, which is only called
when translated strings are actually needed. Rely on
`createTranslator()` to cache the translator between calls.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #817