build: Allow release candidate (rc) versions
After this, running `bump2version rc` will increment the `rc` value if
the current version is an rc. While `bump2version` has a good way to
handle a release process that always goes through a release candidate
phase, it doesn't have good support for an optional release candidate
phase.

For example, if the current version is 6.45.0, `bump2version rc` will
change the version to 6.45.0rc1. That would be wrong as 6.45.0rc1 is
older than 6.45.0. Therefore, rc releases should only be done manually
by developers to ensure the new version is correct. See the README for
details.

Note that the trailing space and real tabs in the config file are
intentional as that matches how `ConfigParser` will write back the file
when bumping `current_version`. Without that we'd just get a spurious
whitespace change on the next version bump.