ci: Perform full fetch for wheel build
`setuptools_scm` needs the git history at least to the most recent tag
or it can't derive the version correctly. Without the history, it
creates a version like `0.1.dev1+g33e4507` and prints a warning like
`"/home/runner/work/kolibri-explore-plugin/kolibri-explore-plugin" is
shallow and may cause errors`. This works when the checkout is directly
on a tag, but it makes the artifacts from other builds have the wrong
version number.