Refactor release workflow
Currently the build steps are designed to handle running a release from
a tag ref so that you can push a tag rather than creating one from the
bump2release workflow. However, the main workflow is only triggered on
pushes to master and stable.

Refactor the release part into a separate workflow that can be either
triggered by a push to a `v*` tag or called from bump2release. The
release workflow in turn runs the main build worflow with a specified
ref.

Helps: #815