Handlers: Fix root URL redirect
The redirect should happen after the loading state is set. Each handler
sets loading to true at the very beginning and to false at the very end.
But because of the redirect, the TOPICS handler (showChannels) was
starting in the middle of the ROOT handler (decideWelcome). Thus the
state was mixed up and not to be trusted.

Fixes #799