build: Copy setup.cfg data to pyproject.toml
This is needed to use newer versions of kolibri-tools, which will read
configuration from `pyproject.toml` in preference to `setup.cfg`, if the
former exists.

See https://github.com/learningequality/kolibri/issues/11146

For now, keep the old configuration present in `setup.cfg` as well, in
case people are using an older version of kolibri-tools. That can be
dropped once we bump our dependency.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>