WIP: build: Pass multiple namespaces to i18n-extract-messages
This speeds up the i18n extraction process, but it cannot remove the
need for the workaround from
https://github.com/endlessm/kolibri-explore-plugin/issues/846 where the
per-namespace CSV files need to be combined into a single JSON file for
webpack to load. Currently that’s done manually.

This commit is still WIP as it needs a kolibri-tools version bump to be
able to rely on the new behaviour, which is currently unmerged upstream
at https://github.com/learningequality/kolibri/pull/11357, and not in a
release yet.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: https://github.com/endlessm/kolibri-explore-plugin/issues/846