views: Use cache_control decorator instead of metadata_cache
The Kolibri `metadata_cache` decorator is tied to the content lifetime,
but that has nothing to do with the channel app data. Furthermore,
`metadata_cache` in Kolibri 0.16.0-beta6 is broken for this usage since
regular `View` classes don't have `render` functions[1].

Instead, simply use the `cache_control` decorator to set the max age for
a week. Since these URLs always contain the plugin version (and the
files in the channel zips are generally cache busting), it should be
safe to essentially never expire the responses.

Fixes: #872

1. https://github.com/learningequality/kolibri/issues/11362