tests: Add fixture to create content directory
The `create_contentdir` function takes the fake channel DB JSON files
and creates a content directory with the channel databases and content
files for testing. That's made available as the `contentdir` pytest
fixture.

A standalone script is provided as a convenience and for exercising the
functionality outside of the test suite.