tests: Add fixtures for channel and content imports
This adds a few pytest fixtures for running channel and content import
jobs. Beyond creating a facility and a facility user with appropriate
permissions, 2 fixtures for handling interactions between Django and
SQLAlchemy. Some of this is lifted from Kolibri and/or depends on
Kolibri internals. We'll see how well they hold up over time.