jobs: Fix remotecontentimport node_ids semantics
As a convenience, the generateed `remotecontentimport` task was setting
`node_ids` to an empty list if it wasn't specified. However, an empty
list and `None` are 2 different cases. An empty list is used when no
node IDs are to be imported, and `None` is used when all node IDs are to
be imported. Even though we're not actually downloading all nodes here,
fix the semantics and require that the thumbnails tasks explicitly
request no nodes. This will be used later in testing.