jobs: Lock database when enqueueing background task
The storage hook depends on the `BackgroundTask` containing the current
job ID. However, it's possible the job will change state and the hook
will run before the job ID has been saved. That would prevent the hook
from properly synchronizing the updated job state. To prevent that, lock
the database while the task is being enqueued until the job ID is saved.