Download: Skip unneeded channel import tasks
Kolibri will always make network requests for a channel import task in
order to check if the remote version is newer than the local version.
However, the starter packs already specify the version we want, so we
can check if that's available and skip creating the channel import task.
Without that, it's not possible to complete the collection downloader
without network access even if all the content is available.

Fixes: #890