tests: Run ContentServer in same process
When the `ContentServer` is run in a separate process with
`multiprocessing`, none of the log messages are recorded. Instead, use a
separate thread in the current process. All that needs to happen:

* `serve_forever()` is run in a separate thread
* `shutdown()` is called from the main thread
* `server_close()` is called to close the socket

Now that the log messages are visible, it's clear that the thread name
in the handler log message is redundant since our pytest default format
includes the thread name.