tests: Also copy channel JSON files into content directory
In order for `ContentServer` to support the
`/api/public/v1/channels/lookup/<channel_id>` endpoint, it needs to be
able to introspect the channel database. Ideally this would open the
actual sqlite database and use Kolibri's routines, but that would be
complicated. Instead, copy the input JSON files into the content
directory so the server can get the channel data from them without using
the databases.