Download: Import extra channels and thumbnails with single task
In order to import an extra channel and its thumbnails, the storage hook
would detect a completed channel import task and then create a thumbnail
import task for that channel dynamically. However, now that channel
import tasks are skipped if the channel is already available, no
completed channel import task would arrive to trigger the thumbnail task
creation.

That's an unlikely scenario since the installation is expected to come
with either no content or fully populated channels, but it can be
handled better. Instead of using 2 tasks, use a single `remoteimport`
task that combines `remotechannelimport` and `remotecontentimport`. The
hook action is kept in case there are existing installations that hadn't
completed the background tasks yet, but hopefully it can be removed
someday.