Add helper functions to check for installed
Improve the iterator for tags and thumbnail jobs. In the same way as
with the other iterators (metadata and content) these can be used to
check if the installation phase was done (when they return an empty
result). Add helper functions too.

These helper functions can be used to set the installation state in the
CollectionState model.